Separate tests of different components into different files
